Numprocs management

+++ SimpleTask.py
@@ -1,27 +1,17 @@
 import subprocess
-from validate import validate_task_config
 import signal
 import time
-from datetime import timedelta
 import os
-from enum 	import Enum, auto
-from _io	import TextIOWrapper
 import logging
+from Task       import Task
+from validate   import validate_task_config
+from datetime   import timedelta
+from _io	    import TextIOWrapper
+from State      import State, RUNNING_STATES, STOPPED_STATES
 
 TICK_RATE = 0.5
 
-class State(Enum):
-    STARTING = auto()
-    RUNNING = auto()
-    BACKOFF = auto()
-    FATAL = auto()
-    EXITED = auto()
-    STOPPED = auto()
-    STOPPING = auto()
-    UNKNOWN = auto()
-    NEVER_STARTED = auto()
-
-class Task:
+class SimpleTask(Task):
     name: str
     cmd: list
     numprocs: int
@@ -102,6 +92,10 @@ class Task:
         return file
 
     def start(self):
+        if self.processus_status in RUNNING_STATES: # If process is already started or in starting progress
+            print(f"{self.name} : ERROR (already started)")
+            return {"success": [], "errors": [self]}
+
         if self.stdout is not None:
             self.stdout_file = self.open_with_umask(self.stdout)
         else:
@@ -114,7 +108,6 @@ class Task:
 
         self.processus_time_start = time.time()
         self.processus_status = State.STARTING
-        logging.info(f"{self.name} starting")
 
         try:
             self.process = subprocess.Popen(
@@ -126,33 +119,19 @@ class Task:
                 env=self.env,
                 start_new_session=True
             )
+            logging.info(f"{self.name} starting")
+            return {"success": [self], "errors": []}
+
         except Exception as e:
-            logging.error(f"Failed to start {self.name}: {e}")
-
-    # def start(self):
-    #     if self.stdout is not None:
-    #         self.stdout_file = self.open_with_umask(self.stdout)
-    #     if self.stderr is not None:
-    #         self.stderr_file = self.open_with_umask(self.stderr)
-
-    #     self.processus_time_start = time.time()
-    #     self.processus_status = State.STARTING
-    #     logging.info(f"{self.name} starting")
-
-    #     try:
-    #         self.process = subprocess.Popen(
-    #             self.cmd,
-    #             stdout=self.stdout_file,
-    #             stderr=self.stderr_file,
-    #             text=True,
-    #             cwd=self.workingdir,
-    #             env=self.env,
-    #             start_new_session=True
-    #         )
-    #     except Exception as e:
-    #         logging.error(f"Failed to start {self.name}: {e}")
-
+            logging.error(f"{self.name} fatal : {e}")
+            self.processus_status = State.FATAL
+            return {"success": [], "errors": [self]}
+    
     def stop(self):
+        if self.processus_status in STOPPED_STATES:
+            print(f"{self.name} : ERROR (not running)") 
+            return {"success": [], "errors": [self]}
+
         signals = {
             "TERM": signal.SIGTERM,
             "KILL": signal.SIGKILL,
@@ -163,6 +142,9 @@ class Task:
             "QUIT": signal.SIGQUIT,
         }
         sig = signals.get(self.stopsignal)
+        if self.process is None:
+            self.processus_status = State.STOPPED
+            return {"success": [], "errors": [self]}
         self.process.send_signal(sig)
         if self.process.poll() is not None:
             self.processus_status = State.STOPPED
@@ -172,54 +154,56 @@ class Task:
             self.processus_time_stop = time.time()
             self.processus_status = State.STOPPING
             logging.info(f"{self.name} stopping")
+        return {"success": [self], "errors": []}
 
-    def supervise(self) -> str:
-        if self.processus_status == State.STARTING:
-            if self.process.poll() is not None and self.process.poll() not in self.exitcodes:
-                if self.retry < self.startretries:
-                    self.retry += 1
-                    self.processus_status = State.BACKOFF
-                    logging.info(f"{self.name} backoff")
+    def supervise(self):
+        if self.process is not None:
+            poll_state = self.process.poll()
+            if self.processus_status == State.STARTING:
+                if poll_state is not None and poll_state not in self.exitcodes:
+                    if self.retry < self.startretries:
+                        self.retry += 1
+                        self.processus_status = State.BACKOFF
+                        logging.info(f"{self.name} backoff")
+                        self.close_redir()
+                    else:
+                        self.processus_status = State.FATAL
+                        logging.info(f"{self.name} fatal")
+                        self.close_redir()
+                elif time.time() - self.processus_time_start >= self.starttime:
+                    self.processus_status = State.RUNNING
+                    logging.info(f"{self.name} running")
+            elif self.processus_status == State.BACKOFF:
+                self.start()
+            elif self.processus_status == State.RUNNING:
+                if poll_state is not None:
                     self.close_redir()
-                else:
-                    self.processus_status = State.FATAL
-                    logging.info(f"{self.name} fatal")
+                    if poll_state in self.exitcodes:
+                        self.processus_status = State.STOPPED
+                        logging.info(f"{self.name} stopped")
+                    elif self.autorestart:
+                        self.retry = 0
+                        self.processus_status = State.BACKOFF
+                        logging.info(f"{self.name} backoff")
+                    else:
+                        self.processus_status = State.FATAL
+                        logging.info(f"{self.name} fatal")
+            elif self.processus_status == State.STOPPING:
+                if poll_state is not None:
                     self.close_redir()
-            elif time.time() - self.processus_time_start >= self.starttime:
-                self.processus_status = State.RUNNING
-                logging.info(f"{self.name} running")
-        elif self.processus_status == State.BACKOFF:
-            time.sleep(0.5)
-            self.start()
-        elif self.processus_status == State.RUNNING:
-            if self.process.poll() is not None:
-                self.close_redir()
-                if self.process.poll() in self.exitcodes:
                     self.processus_status = State.STOPPED
                     logging.info(f"{self.name} stopped")
-                elif self.autorestart:
-                    self.retry = 0
-                    self.processus_status = State.BACKOFF
-                    logging.info(f"{self.name} backoff")
-                else:
-                    self.processus_status = State.FATAL
-                    logging.info(f"{self.name} fatal")
-        elif self.processus_status == State.STOPPING:
-            if self.process.poll() is not None:
-                self.close_redir()
-                self.processus_status = State.STOPPED
-                logging.info(f"{self.name} stopped")
-            elif time.time() - self.processus_time_stop >= self.stoptime:
-                self.process.kill()
-                self.close_redir()
-                self.processus_status = State.STOPPED
-                logging.info(f"{self.name} stopped")
+                elif time.time() - self.processus_time_stop >= self.stoptime:
+                    self.process.kill()
+                    self.close_redir()
+                    self.processus_status = State.STOPPED
+                    logging.info(f"{self.name} stopped")
 
 
     def status(self):
         buffer = f"{self.name:<32}{self.processus_status.name:<10}"
 
-        if self.processus_status == State.RUNNING:
+        if self.processus_status == State.RUNNING and self.process is not None:
             uptime = timedelta(seconds=int(time.time() - self.processus_time_start))
             buffer += f"pid {self.process.pid}, uptime {uptime}"
         if self.processus_status == State.STOPPED:
@@ -228,6 +212,24 @@ class Task:
                 buffer += f"{stop_time}"
             else:
                 buffer += f"Not started"
+        print(buffer)
 
-        return buffer
-        
+    def shutdown(self):
+        if self.processus_status in STOPPED_STATES:
+            return {"success": [], "errors": [self]}
+
+        if self.process is None:
+            self.processus_status = State.STOPPED
+            return {"success": [], "errors": [self]}
+        # Set new stoptime 
+        self.stoptime = 2 
+        self.process.send_signal(signal.SIGTERM)
+        if self.process.poll() is not None:
+            self.processus_status = State.STOPPED
+            logging.info(f"{self.name} shutdown complete")
+            self.close_redir()
+        else:
+            self.processus_time_stop = time.time()
+            self.processus_status = State.STOPPING
+            logging.info(f"{self.name} shutting down")
+        return {"success": [self], "errors": []}
